requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv